+++ packages/components/src/components/as-range-slider/slider/as-range-slider.tsx
@@ -1,4 +1,5 @@
 import { Component, Event, EventEmitter, Prop, State, Watch } from '@stencil/core';
+import getDecimalPlaces from '../../../utils/get-decimal-places';
 import { Thumb } from '../thumb/as-range-slider-thumb';
 
 @Component({
@@ -68,6 +69,7 @@ export class RangeSlider {
 
   /**
    * If this property receives a function, it will be used to format the numbers (eg. for adding $ or €).
+   *
    * @type {function (value: number)}
    * @memberof RangeSlider
    */
@@ -302,6 +304,11 @@ export class RangeSlider {
   }
 
   private _getStepValue(value) {
-    return Math.round(value / this.step) * this.step;
+    const stepValue = (value / this.step) * this.step;
+    return this.roundToStep(stepValue, this.step);
+  }
+
+  private roundToStep(numberToRound: number, step: number) {
+    return Number.parseFloat(numberToRound.toFixed(getDecimalPlaces(step)));
   }
 }