+++ packages/styles/src/switch/test/switch.spec.js
@@ -1,6 +1,0 @@
-const Utils = require('../../test-utils');
-const u = new Utils(__dirname);
-
-module.exports = [
-  u.spec('switch')
-];